requirements.txt
streamlit
python-dotenv
langchain
langchain-openai
langchain-community
langchain-core
openai
langgraph
langsmith
fastapi
uvicorn
pydantic